# internal/apiserver/store/mysql/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package sqlc

import (
	"time"
)

// 用户与 Gateway Agent 的持续对话
type Chat struct {
	// 对话唯一标识
	ID uint64 `json:"id"`
	// 对话创建时间
	CreatedAt time.Time `json:"created_at"`
	// 对话最后活跃时间
	UpdatedAt time.Time `json:"updated_at"`
}

// 对话中的消息
type ChatMessage struct {
	// 消息唯一标识
	ID uint64 `json:"id"`
	// 消息所属对话
	ChatID uint64 `json:"chat_id"`
	// 消息角色
	Role string `json:"role"`
	// 消息正文
	Content string `json:"content"`
	// 消息创建时间
	CreatedAt time.Time `json:"created_at"`
}